requirements.txt
anthropic>=0.38.0,<1.0
streamlit>=1.0
python-dotenv>=1.0
httpx>=0.24.0
